pyproject.toml
[project]
name = "emg2qwerty"
version = "0.1.0"
description = "Baselines for modeling QWERTY typing from surface electromyography."
authors = [{ name = "Viswanath Sivakumar", email = "viswanath@meta.com" }]
requires-python = ">=3.10.13"
readme = "README.md"

dependencies = [
    "click>=8.1.7",
    "datasets>=2.19.2",
    "matplotlib>=3.4.3,<4",
    "seaborn>=0.13.0,<0.14",
    "typer>=0.12.3,<0.13",
    "h5py>=3.11.0",
    "boto3>=1.36.1",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "lightning-bolts>=0.7.0",
    "mne-bids>=0.15.0",
    "nltk>=3.8.1",
    "numpy>=1.24.4",
    "omegaconf>=2.3.0",
    "pandas>=2.0.3",
    "python-levenshtein>=0.12.2",
    "pytorch-lightning>=1.8.6",
    "pyyaml>=5.4.1",
    "pydantic>=2.5.2",
    "scipy>=1.10.1",
    "scikit-learn>=1.3.2",
    "torch>=2.3.0",
    "torchaudio>=2.3.0",
    "torchmetrics>=0.11.4",
    "trio>=0.23.2",
    "unidecode>=1.3.8",
    "camel-kenlm",
]

[dependency-groups]
dev = [
    "pre-commit>=3.7.1",
    "ruff>=0.9.9",
    "pytest>=7.4.4",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=5.0.0",
    "pytest-env>=1.1.3",
    "pytest-mock>=3.14.0",
    "pytest-rerunfailures>=14.0",
    "pytest-xdist[psutil]==3.5.0",
    "hypothesis>=6.97.6",
    "ipykernel>=6.0.1,<7",
    "ipython>=7.25.0,<8",
    "ipdb>=0.13.9,<0.14",
    "mypy>=0.910,<0.911",
    "python-dotenv>=0.21.0,<0.22",
    "pydantic-settings>=2.3.3,<3",
]

[tool.uv]

[tool.uv.sources]
camel-kenlm = { git = "https://github.com/stepanhruda/camel-kenlm.git", rev = "a6e906b2b30497dd999cb9c84c42c5111f8616e0" }

[tool.hatch.build.targets.sdist]
include = ["emg2qwerty"]

[tool.hatch.build.targets.wheel]
include = ["emg2qwerty"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
]
line-length = 100
indent-width = 4
select = ["E", "F", "I"] # E and F for flake8, I for isort

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["emg2qwerty"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E712", "ARG001"]
"research/*.py" = ["T201", "ARG", "E402", "E501", "F841"]

[tool.mypy]
check_untyped_defs = true
follow_imports = "normal"
implicit_optional = false
show_column_numbers = true
show_error_codes = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torch.*"
follow_imports = "skip"
follow_imports_for_stubs = false

[[tool.mypy.overrides]]
module = "pytorch_lightning"
follow_imports = "skip"
follow_imports_for_stubs = false
